large-file: advertise CAP_LARGE_FILES + 64-bit size plumbing

+++ src/protocol.h
@@ -50,15 +50,22 @@ extern void qbuf_add (struct qbuf *q, void *buf, guint32 len);
 struct htlc_conn;
 
 struct htxf_conn {
-    guint32 data_size, data_pos, rsrc_size, rsrc_pos;
-    guint32 total_size, total_pos;
+    /* All size / position fields are guint64 to support the
+	 * Large-File extension (CAP_LARGE_FILES). The protocol still
+	 * defaults to 32-bit on the wire; the upgrade only matters
+	 * when both peers negotiate the capability. Internally we
+	 * always operate in 64-bit so progress reporting, resume,
+	 * and rename-on-collision logic do not have to branch on
+	 * the negotiated mode. */
+    guint64 data_size, data_pos, rsrc_size, rsrc_pos;
+    guint64 total_size, total_pos;
     /* Server's data-fork size from the file listing, captured at
 	 * xfer_new time. Used by xfer_go to choose between resume
 	 * (local exists and is strictly smaller than server) and
 	 * rename-on-collision (local is the same size or larger, or
 	 * server size is unknown). 0 == unknown — listing wasn't
 	 * available for this transfer. */
-    guint32 srv_data_size;
+    guint64 srv_data_size;
     /* Lifecycle: htxf_conn is reference-counted to handle the
 	 * cross-thread ownership knot between the xfers[] array, the
 	 * per-xfer worker thread, and any pending main-thread idles
